Stage all files and commit current work
